Conflicts:
	arch/arm/configs/halaszk_defconfig_N900
	drivers/cpufreq/cpufreq_interactive.c
	drivers/cpufreq/exynos5420-cpufreq.c

+++ drivers/sensorhub/stm/factory/temphumidity_shtc1.c
@@ -209,8 +209,8 @@ err:
 
 static int convert_adc_to_temp(struct ssp_data *data, unsigned int adc)
 {
-	u8 low = 0, mid = 0;
-	u8 high;
+	int low = 0, mid = 0;
+	int high;
 
 	if (!data->cp_thm_adc_table || !data->cp_thm_adc_arr_size) {
 		/* using fake temp */
@@ -260,8 +260,8 @@ static int hub_thm_get_adc(struct ssp_data *data, u32 channel)
 
 static int convert_batt_to_temp(struct ssp_data *data, unsigned int adc)
 {
-	u8 low = 0, mid = 0;
-	u8 high;
+	int low = 0, mid = 0;
+	int high;
 
 	if (!data->batt_thm_adc_table || !data->batt_thm_adc_arr_size) {
 		/* using fake temp */
@@ -284,8 +284,8 @@ static int convert_batt_to_temp(struct ssp_data *data, unsigned int adc)
 
 static int convert_chg_to_temp(struct ssp_data *data, unsigned int adc)
 {
-	u8 low = 0, mid = 0;
-	u8 high;
+	int low = 0, mid = 0;
+	int high;
 
 	if (!data->chg_thm_adc_table || !data->chg_thm_adc_arr_size) {
 		/* using fake temp */
